Created v2 reader, added id3v2 struct and moved logical code from main to reader
