+++ src/models/v2/id3v2.rs
@@ -1,0 +1,10 @@
+use super::super::{Version, v2::{header::*, frames::*}};
+
+#[derive(Debug)]
+pub struct ID3v2 {
+
+    pub version: Version,
+    pub header: Header,
+    pub frames: Vec<Frame>
+
+}
