+++ src/models/v2/mod.rs
@@ -1,2 +1,4 @@
-pub mod frames;
+pub mod id3v2;
+
 pub mod header;
+pub mod frames;
